get cake to kind of work for me
In build.ps1, I changed $PS_SCRIPT_ROOT to $PSScriptRoot, so that the next
two lines have the right variable name.

In build.cake, I commented out (with "//M ") lines that led to a null ref
error when I do '.\build.ps1 -showdescription.

At this point, I'm able to run cake, and it invokes msbuild. However,
there's a build error because System.Attributes is in an assembly that's
not referenced.